Added Scuba Guy Enemy and Began Work On Enemy AI
Right now, the Scuba Guy's AI has 3 planned phases of behavior: idle
(which is moving back and forth in their caves), pursing (chasing the
player), fleeing (trying to leave the level and close out the gameplay).
Idle phase is fully implemented. Working on pursuing phase now. Working
on determining direction to move based upon where the blocking soil is
and where the player is in relation to Scuba Guy.